add logger object and traverse full filesystem path.

+++ api/libretime_api/storage/views/file.py
@@ -1,5 +1,7 @@
+import logging
 import os
 
+from django.conf import settings
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from django.utils.encoding import filepath_to_uri
@@ -7,8 +9,11 @@ from rest_framework import viewsets
 from rest_framework.decorators import action
 from rest_framework.serializers import IntegerField
 
+from ...schedule.models import Schedule
 from ..models import File
 from ..serializers import FileSerializer
+
+logger = logging.getLogger(__name__)
 
 
 class FileViewSet(viewsets.ModelViewSet):
@@ -30,18 +35,23 @@ class FileViewSet(viewsets.ModelViewSet):
         response["X-Accel-Redirect"] = redirect_uri
         return response
 
-    @action(detail=True, methods=["DELETE"])
-    def delete_file(self, request, pk=None):  # pylint: disable=invalid-name
+    def destroy(self, request, *args, **kwargs):  # pylint: disable=invalid-name
+        pk = kwargs.get("pk", None)
         pk = IntegerField().to_internal_value(data=pk)
 
         file = get_object_or_404(File, pk=pk)
-        path = file.filepath
+
+        # Check if the file is scheduled to be played in the future
+        if Schedule.is_file_scheduled_in_the_future(file_id=pk) is True:
+            return HttpResponse(status=409)
+
+        path = os.path.join(settings.CONFIG.storage.path, file.filepath)
 
         try:
             if os.path.isfile(path):
                 os.remove(path)
         except OSError as exception:
-            logger.error(f"Could not delete file from storage: {exception}")
-            return HttpResponse(status=404)
-        file.delete()
+            logger.error("Could not delete file from storage: %s", exception)
+            return HttpResponse(status=500)
+        self.perform_destroy(file)
         return HttpResponse(status=204)
